Implement proper download using chrome.experimental.downloads API
BTW I refactored a bit the request handling code and added the
option to open tabs in a new window with Shift+Enter.
I also resorted to use the chrome.tabs API to open new tabs instead
of clicking elements, which was a really nasty hack we used to
simulate downloads.
When opening a download with the new API, Chrome doesn't ask to
keep the file if it may be dangerous so we open a new tab pointing
to the download manager.